refactor: Scaffold Component에서 분리
