+++ app/src/main/java/nextstep/signup/ui/SignUpScreen.kt
@@ -2,20 +2,12 @@ package nextstep.signup.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -24,7 +16,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
-import kotlinx.coroutines.launch
 import nextstep.signup.R
 import nextstep.signup.ui.component.SignUpButton
 import nextstep.signup.ui.component.SignUpHeaderText
@@ -36,133 +27,123 @@ import nextstep.signup.ui.model.SignUpInformation
 import nextstep.signup.ui.model.UserName
 
 @Composable
-fun SignUpScreen() {
-    val coroutineScope = rememberCoroutineScope()
-    val snackbarHostState = remember { SnackbarHostState() }
-    val snackbarMessage = stringResource(R.string.sign_up_completed)
+fun SignUpScreen(
+    modifier: Modifier = Modifier,
+    snackBarEvent: (String) -> Unit = {},
+) {
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        var userName by rememberSaveable { mutableStateOf(UserName()) }
+        var email by rememberSaveable { mutableStateOf(Email()) }
+        var password by rememberSaveable { mutableStateOf(Password()) }
+        var passwordConfirm by rememberSaveable { mutableStateOf(PasswordConfirm()) }
 
-    Scaffold(
-        modifier = Modifier.fillMaxSize(),
-        snackbarHost = { SnackbarHost(snackbarHostState) },
-    ) { paddingValues ->
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxSize()
-                    .padding(paddingValues)
-                    .verticalScroll(rememberScrollState()),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            var userName by rememberSaveable { mutableStateOf(UserName()) }
-            var email by rememberSaveable { mutableStateOf(Email()) }
-            var password by rememberSaveable { mutableStateOf(Password()) }
-            var passwordConfirm by rememberSaveable { mutableStateOf(PasswordConfirm()) }
-
-            val signUpInformation =
-                SignUpInformation(
-                    userName = userName,
-                    email = email,
-                    password = password,
-                    passwordConfirm = passwordConfirm,
-                )
-
-            Spacer(modifier = Modifier.padding(top = 84.dp))
-
-            SignUpHeaderText(
-                modifier =
-                    Modifier.padding(
-                        start = 28.dp,
-                        end = 28.dp,
-                    ),
-                text = stringResource(R.string.sign_up_header),
+        val signUpInformation =
+            SignUpInformation(
+                userName = userName,
+                email = email,
+                password = password,
+                passwordConfirm = passwordConfirm,
             )
 
-            Column(
+        val snackbarMessage = stringResource(R.string.sign_up_completed)
+
+        Spacer(modifier = Modifier.padding(top = 84.dp))
+
+        SignUpHeaderText(
+            modifier =
+                Modifier.padding(
+                    start = 28.dp,
+                    end = 28.dp,
+                ),
+            text = stringResource(R.string.sign_up_header),
+        )
+
+        Column(
+            modifier =
+                Modifier.padding(
+                    start = 28.dp,
+                    end = 28.dp,
+                    top = 42.dp,
+                    bottom = 42.dp,
+                ),
+        ) {
+            SignUpTextField(
                 modifier =
-                    Modifier.padding(
-                        start = 28.dp,
-                        end = 28.dp,
-                        top = 42.dp,
-                        bottom = 42.dp,
-                    ),
-            ) {
-                SignUpTextField(
-                    modifier =
-                        Modifier.fillMaxWidth(),
-                    label = stringResource(R.string.username_label),
-                    value = userName.text,
-                    onValueChange = { value ->
-                        userName = userName.copy(text = value)
-                    },
-                    isValid = userName.isValid(),
-                    errorMessage = userName.errorMessage(),
-                )
+                    Modifier.fillMaxWidth(),
+                label = stringResource(R.string.username_label),
+                value = userName.text,
+                onValueChange = { value ->
+                    userName = userName.copy(text = value)
+                },
+                isValid = userName.isValid(),
+                errorMessage = userName.errorMessage(),
+            )
 
-                SignUpTextField(
-                    modifier =
-                        Modifier
-                            .fillMaxWidth()
-                            .padding(top = 36.dp),
-                    label = stringResource(R.string.email_label),
-                    value = email.text,
-                    onValueChange = { value ->
-                        email = email.copy(text = value)
-                    },
-                    isValid = email.isValid(),
-                    errorMessage = email.errorMessage(),
-                    keyboardType = KeyboardType.Email,
-                )
-
-                SignUpTextField(
-                    modifier =
-                        Modifier
-                            .fillMaxWidth()
-                            .padding(top = 36.dp),
-                    label = stringResource(R.string.password_label),
-                    value = password.text,
-                    onValueChange = { value ->
-                        password = password.copy(text = value)
-                        passwordConfirm = passwordConfirm.copy(passwordText = value)
-                    },
-                    isValid = password.isValid(),
-                    errorMessage = password.errorMessage(),
-                    keyboardType = KeyboardType.Password,
-                    visualTransformation = PasswordVisualTransformation(),
-                )
-
-                SignUpTextField(
-                    modifier =
-                        Modifier
-                            .fillMaxWidth()
-                            .padding(top = 36.dp),
-                    label = stringResource(R.string.password_confirm_label),
-                    value = passwordConfirm.text,
-                    onValueChange = { value ->
-                        passwordConfirm = passwordConfirm.copy(text = value)
-                    },
-                    isValid = passwordConfirm.isValid(),
-                    errorMessage = passwordConfirm.errorMessage(),
-                    keyboardType = KeyboardType.Password,
-                    visualTransformation = PasswordVisualTransformation(),
-                )
-            }
-
-            SignUpButton(
+            SignUpTextField(
                 modifier =
                     Modifier
                         .fillMaxWidth()
-                        .height(50.dp)
-                        .padding(start = 32.dp, end = 32.dp),
-                text = stringResource(R.string.sign_up_button_label),
-                onclick = {
-                    coroutineScope.launch {
-                        snackbarHostState.showSnackbar(snackbarMessage)
-                    }
+                        .padding(top = 36.dp),
+                label = stringResource(R.string.email_label),
+                value = email.text,
+                onValueChange = { value ->
+                    email = email.copy(text = value)
                 },
-                enabled = signUpInformation.isEligibleForSignUp(),
+                isValid = email.isValid(),
+                errorMessage = email.errorMessage(),
+                keyboardType = KeyboardType.Email,
             )
 
-            Spacer(modifier = Modifier.padding(bottom = 28.dp))
+            SignUpTextField(
+                modifier =
+                    Modifier
+                        .fillMaxWidth()
+                        .padding(top = 36.dp),
+                label = stringResource(R.string.password_label),
+                value = password.text,
+                onValueChange = { value ->
+                    password = password.copy(text = value)
+                    passwordConfirm = passwordConfirm.copy(passwordText = value)
+                },
+                isValid = password.isValid(),
+                errorMessage = password.errorMessage(),
+                keyboardType = KeyboardType.Password,
+                visualTransformation = PasswordVisualTransformation(),
+            )
+
+            SignUpTextField(
+                modifier =
+                    Modifier
+                        .fillMaxWidth()
+                        .padding(top = 36.dp),
+                label = stringResource(R.string.password_confirm_label),
+                value = passwordConfirm.text,
+                onValueChange = { value ->
+                    passwordConfirm = passwordConfirm.copy(text = value)
+                },
+                isValid = passwordConfirm.isValid(),
+                errorMessage = passwordConfirm.errorMessage(),
+                keyboardType = KeyboardType.Password,
+                visualTransformation = PasswordVisualTransformation(),
+            )
         }
+
+        SignUpButton(
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .height(50.dp)
+                    .padding(start = 32.dp, end = 32.dp),
+            text = stringResource(R.string.sign_up_button_label),
+            onclick = {
+                snackBarEvent(snackbarMessage)
+            },
+            enabled = signUpInformation.isEligibleForSignUp(),
+        )
+
+        Spacer(modifier = Modifier.padding(bottom = 28.dp))
     }
 }
